requirements.txt
absl-py==0.13.0
actionlib==1.13.2
alabaster==0.7.12
angles==1.9.13
appdirs==1.4.3
apturl==0.5.2
astunparse==1.6.3
atari-py==0.2.9
atomicwrites==1.1.5
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
Babel==2.10.3
backcall==0.1.0
backports.entry-points-selectable==1.1.0
bagpy==0.4.10
base-local-planner==1.17.1
bcrypt==3.1.7
beautifulsoup4==4.8.2
bitstring==3.1.9
bleach==3.1.1
blessings==1.6
blinker==1.4
bondpy==1.8.6
box2d-py==2.3.8
breezy==3.0.2
Brlapi==0.7.0
cachetools==4.2.2
camera-calibration==1.15.3
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==0.4.23
catkin-pkg-modules==0.4.23
cbor==1.0.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.0.4
chrome-gnome-shell==0.0.0
Click==7.0
cloudpickle==1.6.0
colorama==0.4.3
command-not-found==0.3
commonmark==0.9.1
configobj==5.0.6
constantly==15.1.0
controller-manager==0.19.5
controller-manager-msgs==0.19.5
cryptography==2.8
cssselect==1.1.0
cupshelpers==1.0
cv-bridge==1.15.0
cycler==0.10.0
Cython==0.29.14
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
Deprecated==1.2.7
diagnostic-analysis==1.10.4
diagnostic-common-diagnostics==1.10.4
diagnostic-updater==1.10.4
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
docutils==0.16
dulwich==0.19.15
duplicity==0.8.12.0
dynamic-reconfigure==1.7.1
empy==3.3.2
entrypoints==0.3
et-xmlfile==1.0.1
fasteners==0.14.1
fastimport==0.9.8
filelock==3.0.12
flatbuffers==2.0
future==0.18.2
gast==0.5.3
gazebo-plugins==2.9.2
gazebo-ros==2.9.2
gencpp==0.6.5
geneus==3.0.0
genlisp==0.4.18
genmsg==0.5.16
gennodejs==2.0.2
genpy==0.6.15
ghp-import==2.1.0
gnupg==2.3.1
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
gpg===1.13.1-unknown
gpustat==0.6.0
grpcio==1.39.0
gym==0.18.3
h5py==3.6.0
html5lib==1.0.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
image-geometry==1.15.0
imagesize==1.4.1
importlab==0.6.1
importlib-metadata==1.5.0
incremental==16.10.1
iniconfig==1.1.1
interactive-markers==1.12.0
iotop==0.6
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
jdcal==1.0
jedi==0.15.2
Jinja2==2.10.1
joint-state-publisher==1.15.0
joint-state-publisher-gui==1.15.0
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
kdl-parser-py==1.14.1
keras==2.8.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
laser-geometry==1.6.7
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==13.0.0
libcst==0.3.21
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
lz4==3.0.2+dfsg
m2r2==0.3.2
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
Markups==3.0.0
MarkupSafe==1.1.0
matplotlib==3.1.2
mavros==1.9.0
mergedeep==1.3.4
message-filters==1.15.11
mistune==0.8.4
mkdocs==1.3.1
monotonic==1.5
more-itertools==4.2.0
mpi4py==3.0.3
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat==5.0.4
netifaces==0.10.4
networkx==2.6.3
ninja==1.10.2.2
nose==1.3.7
notebook==6.0.3
numexpr==2.7.1
numpy==1.17.4
nvidia-ml-py==11.450.51
nvidia-ml-py3==7.352.0
nvitop==0.5.4
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.3.56
openpyxl==3.0.3
opt-einsum==3.3.0
packaging==20.3
pandas==0.25.3
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.5.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
platformdirs==2.2.0
pluggy==0.13.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.6.1
psutil==5.5.1
py==1.8.1
py-ubjson==0.14.0
py3rosmsgs==1.18.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycrypto==2.6.1
pycryptodomex==3.6.1
pycups==1.9.73
pydash==5.1.0
pydot==1.4.1
pyenchant==2.0.0
PyGithub==1.43.7
pyglet==1.5.15
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymdown-extensions==9.5
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.5
pytest==4.6.9
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-gitlab==2.0.1
python-gnupg==0.4.5
python-markdown-math==0.6
python-qt-binding==0.4.4
python-snappy==0.5.3
PyTrie==0.2
pytype==2021.10.11
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyyaml-env-tag==0.1
pyzmq==18.1.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
recommonmark==0.7.1
regex==2019.8.19
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
resource-retriever==1.12.6
rinoh-typeface-dejavuserif==0.1.3
rinoh-typeface-texgyrecursor==0.1.1
rinoh-typeface-texgyreheros==0.1.1
rinoh-typeface-texgyrepagella==0.1.1
rinohtype==0.5.4
roman==2.0.0
rosbag==1.15.11
rosbag-pandas==0.5.4
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosdep==0.21.0
rosdep-modules==0.21.0
rosdistro==0.8.3
rosdistro-modules==0.8.3
rosgraph==1.15.11
rosinstall==0.7.8
rosinstall-generator==0.1.22
roslaunch==1.15.11
roslib==1.15.8
roslint==0.12.0
roslz4==1.15.11
rosmake==1.15.8
rosmaster==1.15.11
rosmsg==1.15.11
rosnode==1.15.11
rosparam==1.15.11
rospkg==1.3.0
rospkg-modules==1.3.0
rospy==1.15.11
rospy-message-converter==0.5.7
rosservice==1.15.11
rostest==1.15.11
rostopic==1.15.11
rosunit==1.15.8
roswtf==1.15.11
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.11
rqt-dep==0.4.12
rqt-graph==0.4.14
rqt-gui==0.5.2
rqt-gui-py==0.5.2
rqt-image-view==0.4.16
rqt-launch==0.4.9
rqt-logger-level==0.4.11
rqt-moveit==0.5.10
rqt-msg==0.4.10
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.11
rqt-publisher==0.4.10
rqt-py-common==0.5.2
rqt-py-console==0.4.10
rqt-reconfigure==0.5.4
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.13
rqt-robot-steering==0.5.12
rqt-runtime-monitor==0.5.9
rqt-rviz==0.7.0
rqt-service-caller==0.4.10
rqt-shell==0.4.11
rqt-srv==0.4.9
rqt-tf-tree==0.6.2
rqt-top==0.4.10
rqt-topic==0.4.12
rqt-web==0.4.10
rsa==4.7.2
rviz==1.14.8
sb3-contrib==1.2.0
scipy==1.3.3
scour==0.37
screen-resolution-extra==0.0.0
seaborn==0.11.2
SecretStorage==2.3.1
Send2Trash==1.5.0
sensor-msgs==1.13.1
service-identity==18.1.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
smach==2.5.0
smach-ros==2.5.0
smclib==1.8.6
snowballstemmer==2.2.0
sound-play==0.3.12
soupsieve==1.9.5
Sphinx==3.2.1
sphinx-autodoc-typehints==1.4.0
sphinx-bootstrap-theme==0.8.1
sphinx-markdown-parser==0.2.4
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
ssh-import-id==5.10
stable-baselines3==1.2.0
systemd-python==234
tables==3.6.1
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.4
textile==4.0.1
tf==1.13.2
tf-conversions==1.13.2
tf-estimator-nightly==2.8.0.dev2021122109
tf2-geometry-msgs==0.7.5
tf2-kdl==0.7.5
tf2-py==0.7.5
tf2-ros==0.7.5
tf2-sensor-msgs==0.7.5
toml==0.10.0
topic-tools==1.15.11
torch==1.10.2+cu113
torchaudio==0.10.2+cu113
torchvision==0.11.3+cu113
tornado==5.1.1
traitlets==4.3.3
Twisted==18.9.0
txaio==2.10.0
typed-ast==1.4.3
typing-extensions==3.10.0.0
typing-inspect==0.7.1
u-msgpack-python==2.1
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unify==0.5
untokenize==0.1.1
urdfdom-py==0.4.5
urllib3==1.25.8
usb-creator==0.3.7
vcstool==0.2.15
vcstools==0.1.42
virtualenv==20.0.17
wadllib==1.3.3
watchdog==2.1.9
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==2.0.0
wrapt==1.11.2
wsaccel==0.6.2
wstool==0.1.18
xacro==1.14.8
xkit==0.0.0
xlrd==1.1.0
xlwt==1.3.0
yapf==0.32.0
ytsphinx==1.2.1.dev20200430
zipp==1.0.0
zope.interface==4.7.1
